ent-est-v3/services/exams/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
cassandra-driver==3.29.0
minio==7.2.5
python-multipart==0.0.9
pydantic==2.6.4
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
pika==1.3.2
requests==2.31.0
